Separate @parcel/lsp-protocol
