+++ .vscode/launch.json
@@ -12,7 +12,7 @@
       "outFiles": [
         "${workspaceFolder}/packages/utils/parcelforvscode/out/**/*.js"
       ],
-      "preLaunchTask": "npm: watch - packages/utils/parcelforvscode",
+      "preLaunchTask": "Watch VSCode Extension",
       "request": "launch",
       "type": "extensionHost"
     }
