+++ packages/utils/parcel-lsp-protocol/index.ts
@@ -1,0 +1,1 @@
+export * from './src/index';
