+++ packages/utils/parcel-lsp-protocol/tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "module": "commonjs",
     "target": "es6",
-    "outDir": "out",
+    "outDir": "lib",
     "lib": ["es6"],
     "sourceMap": true,
     "rootDir": "src",
