+++ packages/utils/parcelforvscode/package.json
@@ -52,6 +52,7 @@
     "lint": "eslint src --ext ts"
   },
   "devDependencies": {
+    "@parcel/lsp": "2.8.2",
     "@types/glob": "^7.1.3",
     "@types/mocha": "^8.0.4",
     "@types/node": "^12.11.7",
@@ -65,7 +66,7 @@
     "vscode-test": "^1.5.0"
   },
   "dependencies": {
-    "@parcel/lsp": "2.8.2",
+    "@parcel/lsp-protocol": "2.8.2",
     "vscode-languageclient": "^8.0.2"
   }
 }
