+++ packages/utils/parcelforvscode/tsconfig.json
@@ -6,11 +6,9 @@
     "lib": ["es6"],
     "sourceMap": true,
     "rootDir": "src",
-    "strict": true /* enable all strict type-checking options */
-    /* Additional Checks */
-    // "noImplicitReturns": true, /* Report error when not all code paths in function return a value. */
-    // "noFallthroughCasesInSwitch": true, /* Report errors for fallthrough cases in switch statement. */
-    // "noUnusedParameters": true,  /* Report errors on unused parameters. */
+    "strict": true,
+    "allowJs": true
   },
+  "include": ["src/*"],
   "exclude": ["node_modules", ".vscode-test"]
 }
